% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karim.R
\name{copyDir}
\alias{copyDir}
\title{copyDir}
\usage{
copyDir(from, to, pattern = "*")
}
\arguments{
\item{fromDir}{\code{character} a path to the image data}

\item{toProjDir}{\code{character} a path to the projRootDir}
}
\description{
copyDir copy all image data to the corresponding folder
}
