% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExif.R
\name{extractExifr}
\alias{extractExifr}
\title{Select UAV Images Information}
\usage{
extractExifr(path)
}
\arguments{
\item{path}{path to the images files}
}
\value{
data.frame of image positions and travel distance
}
\description{
extract all and returns specific exif information from a list of images
}
