% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalBoundingBox.R
\name{minBB}
\alias{minBB}
\title{Rectangle flight area around points}
\usage{
minBB(points, buffer = 0, epsg = 25832)
}
\arguments{
\item{points}{a sf object, points you want to fly over}

\item{buffer}{buffer distance between the points and the rectangle; defaults 0}

\item{epsg}{reference system}
}
\value{
SpatialPoints: Corners of the flight area
}
\description{
Creates optimal rectangle area around points
}
\details{
The code is based on a Rotating Caliper Algorithm and mostly copy and pasted (see reference)
}
\references{
http://dwoll.de/rexrepos/posts/diagBounding.html
}
\author{
Marvin Ludwig
}
