% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picapos.R
\name{picapos}
\alias{picapos}
\title{picapos take  pictures at a given position}
\usage{
picapos(projectDir = "~", locationName = "autoflightcontrol",
  missionTrackList = NULL, launchPos = NULL, demFn = NULL,
  flightAltitude = 75, climbDist = 7.5, aboveTreeAlt = 15,
  circleRadius = 5, takeOffAlt = 50, presetFlightTask = "remote",
  maxSpeed = 25, followSurfaceRes = 5, altFilter = 1, maxFL = 10,
  windCondition = 1, rcRange = -9999, launchAltitude = -9999,
  uavType = "solo")
}
\arguments{
\item{projectDir}{path to the main folder where several projects can be hosted It will overwrite the DEM based estimation if any other value than -9999}

\item{locationName}{base string for mission filenames}

\item{missionTrackList}{mission tracklist}

\item{launchPos}{launch position}

\item{demFn}{filname of the corresponding DEM data file}

\item{flightAltitude}{set the default flight altitude of the mission. It is 
  assumed that the UAV is started at the highest point of the surveyArea 
  otherwise you have to defined the position of launching.
By default it is set to (\code{= 0.0}). If set to \code{-99} it will be 
calculated from the swath width of the pictures. NOTE This makes only sense for 
\code{followingTerrain = TRUE} to smooth curves.
For \code{flightPlanMode = "waypoint"} camera actions are DISABLED during curve flights.}

\item{climbDist}{climbdist}

\item{aboveTreeAlt}{altitude above trees}

\item{circleRadius}{radius to circle around}

\item{takeOffAlt}{altitude of take off point}

\item{presetFlightTask}{(DJI only) strongly recommended to use "remote" 
       \cr
 Options are: 
\code{"simple_ortho"} takes one picture/waypoint, 
\code{"multi_ortho"} takes 4 picture at a waypoint, two vertically down and two in forward and backward viewing direction and an angele of -60deg,
\code{"simple_pano"} takes a 360 deg panorama picture and 
\code{"remote"} which assumes that the camera is controlled by the remote control (RC)}

\item{maxSpeed}{cruising speed}

\item{followSurfaceRes}{followSurfaceRes}

\item{altFilter}{allowed altitude differences}

\item{maxFL}{maxFL}

\item{windCondition}{1= calm 2= light air 1-5km/h, 3= light breeze 6-11km/h, 4=gentle breeze 12-19km/h 5= moderate breeze 20-28km/h}

\item{rcRange}{range of estimated range of remote control}

\item{launchAltitude}{altitude of launch position}

\item{uavType}{type of uav. currently "djip3" and "solo" are supported}
}
\description{
picapos generates a flight track chaining up point objects with respect to a heterogenous Surface and known obstacles for taking top down pictures. Basic idea is to fly in a serie to object positions with respect to the surface model for taking high resolution pics
}
\examples{
\dontrun{
requires(mapview)
t3 <- picapos(projectDir ="/home/creu/uav/bayerwald",
              locationName = "filzmoosTree",
              missionTrackList="~/uav/bayerwald/Selected_trees_Filz.txt",
              demFn = "~/uav/grossfilz/grosserfilz.tif",
              windCondition = 2,
              uavType = "djip3",
              followSurfaceRes=5,
              launchPos = c(13.409114897133804,48.92039612988935))

mapview(t3$wp,zcol = "altitude",lwd=1,cex=5)+mapview(t3$lp,color="red",cex=5)
}

}
