% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{qgc_survey}
\alias{qgc_survey}
\title{Flight area planning Qgroundcontrol survey data}
\source{
\code{Faculty of Geography Marburg}
}
\description{
Flight area planning example data as typically needed for planning an autonomous survey flight task. The task is planned with the QGroundcontrol survey tool.
}
\details{
Flight area planning Qgroundcontrol survey data
}
\keyword{datasets}
