% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soloTB.R
\name{soloLog}
\alias{soloLog}
\title{Download, reorganize and export the log files from 3DR Solo and radio control unit}
\usage{
soloLog(logFiles = "*.BIN*", logSource = "pixhawk", logDir = tmpDir(),
  downloadOnly = FALSE, netWarn = TRUE, organize = TRUE, makeSP = FALSE)
}
\arguments{
\item{logFiles}{pattern of which kind of logs should be downloaded, default is "*.BIN" for all log files from the Pixhawk. If you want the telemetry logfiles from the radio control choose  "solo.t*" which means all log files...}

\item{logSource}{defines if the logfiles are downloaded from the rc controller or the pixhawk, default is "pixhawk" for loading the files from the pixhawk, for the recent telemetry data files from the rc-controller choose "rc"}

\item{logDir}{(existing) destination path to which the logs should be downloaded to}

\item{downloadOnly}{default = FALSE, set to TRUE  if you ONLY want to download the log files from the solo controller to the logDir}

\item{netWarn}{if true warns and waits before starting a connection to the controller to connect to the solo wifi}

\item{organize}{renames the log and gpx files according to their timeperiod}

\item{makeSP}{generates SP objects from the gpx files}
}
\description{
Wraps the mavtogpx.py converter as provided by the dronkit library. It downloads and/ or converts the 3DR Solo logfiles. Otionally you may import the geometries and data as sp objects in R
}
\note{
for using the solo stuff you need to install: \cr 
sudo pip install pymavlink  \cr 
sudo pip install dronekit-sitl \cr 
sudo pip install dronekit \cr 
sudo apt-get install sshpass \cr 
\cr Additionally you need to be connected to a running 3DR Solo uav
}
\examples{
\dontrun{
## download current telemetry log file from controller and convert it to gpx
soloLog(logFiles = "solo.tlog")

## download all available telemetry logfiles from the controller
soloLog()

## download ALL logfiles from the controller
soloLog(logFiles = "*")
}
}
