% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{tutdata_dem}
\alias{tutdata_dem}
\title{DEM data set of Marburg-Biedenkopf}
\format{
\code{"raster::raster"}
}
\source{
\code{Faculty of Geography UAV derived data from Marburg University Forest first campaign}
}
\description{
DEM data set resampled to 20 m resolution
}
\details{
DEM data set of Marburg-Biedenkopf
}
\keyword{datasets}
