% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{tutdata_dji}
\alias{tutdata_dji}
\title{DJI image of a survey flight}
\format{
\code{"raster::raster"}
}
\source{
\verb{Faculty of Geography UAV derived data from Marburg University Forest first campaign}
}
\description{
DJI image of a survey flight
}
\details{
DJI image of a survey flight
}
\keyword{datasets}
