% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{tutdata_flightarea}
\alias{tutdata_flightarea}
\title{Flight area planning example data}
\source{
\verb{Faculty of Geography Marburg}
}
\description{
Flight area planning example data as typically needed for planning an autonomous survey flight task
}
\details{
Flight area planning example data
}
\keyword{datasets}
