% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{tutdata_flighttrack}
\alias{tutdata_flighttrack}
\title{GPX example data}
\source{
\code{Faculty of Geography UAV derived data from Marburg University Forest first campaign}
}
\description{
GPX example data as derived by a 3DR Solo flight
}
\details{
GPX example data
}
\keyword{datasets}
