% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{tutdata_position}
\alias{tutdata_position}
\title{position example data}
\source{
\verb{Faculty of Geography UAV derived data from Marburg University Forest first campaign}
}
\description{
position data for planning a single flight task with focus on known objects
}
\details{
Virtual object position  coordinates example data
}
\keyword{datasets}
