% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{tutdata_qgc_survey30m}
\alias{tutdata_qgc_survey30m}
\title{Flight area planning Qgroundcontrol planning file for a 30m follow terrain survey flight with the DJI Air 2S}
\source{
\verb{Faculty of Geography Marburg}
}
\description{
Flight area planning example data as typically needed for planning an autonomous survey flight task. The task is planned with the QGroundcontrol survey tool.
}
\details{
Flight area planning Qgroundcontrol survey data 30 m AGL
}
\keyword{datasets}
