% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{data}
\name{tutdata_waypoints}
\alias{tutdata_waypoints}
\title{MAVLINK waypoint example data}
\source{
\verb{Faculty of Geography UAV derived data from Marburg University Forest first campaign}
}
\description{
Waypoint file
}
\details{
MAVLINK waypoint example data
}
\keyword{datasets}
