% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{package}
\name{uavRmp}
\alias{uavRmp}
\alias{uavRmp-package}
\title{UAV R Based Mission Planner}
\description{
The package provides some mission planning functionality for dealing #' with Unmanned Aerial Vehicles
}
\details{
The package provides some mission planning functionality for dealing with Unmanned Aerial Vehicles
}
\note{
It is important to keep in mind that all binaries need to be installed 
correctly on your system.
}
\author{
Chris Reudenbach Lars Opgenoorth Sebastian Richter
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{package}
