% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karim.R
\name{xyz2tif}
\alias{xyz2tif}
\title{Read and Convert xyz DEM/DSM Data as typically provided by the public authorities}
\usage{
xyz2tif(xyzFN = NULL, epsgCode = "25832")
}
\arguments{
\item{xyzFN}{ASCII tect file with xyz values}

\item{epsgCode}{EPSG Code default is  "25832"}
}
\description{
Read xyz data and generate a raster  \code{Raster*} object.
}
\examples{
\dontrun{
# get some typical data as provided by the public authority of Bavaria
setwd(tempdir())
url<-"http://www.ldbv.bayern.de/file/zip/10430/DGM_1_ascii.zip"
res <- curl::curl_download(url, "testdata.zip")
unzip("testdata.zip",junkpaths = TRUE,overwrite = TRUE)
# convert it 
xyz2tif("DGM_1.g01dgm","25832")
# plot it
r<-raster::raster("DGM_1.tif")
raster::plot(r)
} 
}
