% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karim.R
\name{copyDir}
\alias{copyDir}
\title{copyDir}
\usage{
copyDir(fromDir, toProjDir, pattern = "*")
}
\arguments{
\item{fromDir}{\code{character} a path to the image data}

\item{toProjDir}{\code{character} a path to the projRootDir}

\item{pattern}{\code{character} a string pattern for filtering  file list}
}
\description{
copyDir copy all image data to the corresponding folder
}
