% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karim.R
\name{getGPSAltDiff}
\alias{getGPSAltDiff}
\title{getGPSAltDiff}
\usage{
getGPSAltDiff(picPath, demPath)
}
\arguments{
\item{picPath}{\code{character} a path to the image}

\item{demPath}{\code{character} a path to the DEM}
}
\description{
extract the difference of the exif GPS aLtitude tag and a given DEM altitude at this point
}
\examples{
\dontrun{
demFN <- system.file("extdata", "mrbiko.tif", package = "uavRmp")
picFN <- system.file("extdata", "dji.jpg", package = "uavRmp")

}
}
\keyword{internal}
