% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karim.R
\name{maxpos_on_line}
\alias{maxpos_on_line}
\title{applies a line to a raster and returns the position of the maximum value}
\usage{
maxpos_on_line(dem, line)
}
\arguments{
\item{dem}{raster object}

\item{line}{sp object}
}
\description{
applies a line to a raster and returns the position of the maximum value
}
