% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karim.R
\name{maxpos_on_line}
\alias{maxpos_on_line}
\title{applies a line to a raster and returns the position of the maximum value}
\usage{
maxpos_on_line(dem, line)
}
\arguments{
\item{dem}{raster object}

\item{line}{sp object}
}
\description{
applies a line to a raster and returns the position of the maximum value
}
\examples{
## load DEM/DSM 
dem <- raster::raster(system.file("extdata", "mrbiko.tif", package = "uavRmp"))

## generate extraction line object
line <- sp_line(c(8.66821,8.68212),c(50.83939,50.83267),ID="Highest Position",runDir=runDir)

## extract highest position
maxpos_on_line(dem,line)  

}
