% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessingTools.R
\name{selectImages}
\alias{selectImages}
\title{extract exif information}
\usage{
selectImages(path = NULL, position = TRUE, distance = 30,
  image_format = "jpg")
}
\arguments{
\item{path}{path to the images files}

\item{position}{default is TRUE the files are selected by their relative position}

\item{distance}{default is 30 m distance of the image centres}

\item{image_format}{jpg or raw}
}
\description{
extract exif
}
\examples{
wp <- system.file("extdata", "MAVLINK_waypoints.txt", package = "uavRmp")
\dontrun{
solo_upload( missionFile = wp)
}
}
