% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soloTB.R
\name{soloLog}
\alias{soloLog}
\title{Download, reorganize and export the log files from 3DR Solo and radio control unit}
\usage{
soloLog(logFiles = "recent", logSource = "rc", logDir = tmpDir(),
  downloadOnly = FALSE, netWarn = TRUE, organize = TRUE, makeSP = FALSE)
}
\arguments{
\item{logFiles}{If logSource = "pixhawk" the argument "recent" will download the most recent logfile from the pixhawk, other options are "all" for all logfiles or a plain number i.e. "1" for a specific one.}

\item{logSource}{\code{rc} = logfiles from the radio control, \code{pixhawk} = logfiles from the flightcontroller, default is \code{rc}. The radio control is providing the last ten telemetry data files, while the flight controller provides the latest 50 binary logfiles.}

\item{logDir}{(existing) destination path to which the logs should be downloaded to}

\item{downloadOnly}{default = TRUE, set to FALSE  if you want to convert the log files from the solo remote controller to GPX files}

\item{netWarn}{if true warns and waits before starting a connection to the controller to connect to the solo wifi}

\item{organize}{renames the log and gpx files according to their timeperiod}

\item{makeSP}{generates SP objects from the gpx files}
}
\description{
Wraps the mavtogpx.py converter as provided by the dronkit library. It downloads and/ or converts the 3DR Solo logfiles. Otionally you may import the geometries and data as sp objects in R
}
\note{
for using the solo stuff you need to install: \cr 
sudo pip install pymavlink  \cr 
sudo pip install dronekit-sitl \cr 
sudo pip install dronekit \cr 
sudo apt-get install sshpass \cr 
\cr Additionally you need to be connected to a running 3DR Solo uav
}
\examples{
\dontrun{
## download recent telemetry log file from controller and convert it to gpx
soloLog(logFiles = "solo.tlog")

## download the last available logfile from the radio control
soloLog()

## download ALL logfiles from the radio control
soloLog(logFiles = "all")

## download ALL telemetry logfiles from the flight controller
soloLog(logSource = "pixhawk",logFiles = "all")

## download telementry logfile number 5  from the remote control
soloLog(logSource = "rc",logFiles = "5")
}
}
