% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karim.R
\name{sp_point}
\alias{sp_point}
\title{create an spatialpointobject from 1 points}
\usage{
sp_point(
  lon,
  lat,
  ID = "point",
  proj4 = "+proj=longlat +datum=WGS84 +no_defs",
  export = FALSE,
  runDir = runDir
)
}
\arguments{
\item{lon}{lon}

\item{lat}{lat}

\item{ID}{name of point}

\item{proj4}{projection}

\item{export}{write shafefile default = F}

\item{runDir}{\code{character} runtime folder}
}
\description{
create an spatial point object from 1 point and optionally export it as a shapefile
}
\examples{
## creating sp spatial point object
point <- sp_point(8.770362,50.815240,ID="Faculty of Geographie Marburg")

## plot it
raster::plot(point)

}
