% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRst-package.R
\docType{package}
\name{uavRmp}
\alias{uavRmp}
\alias{uavRmp-package}
\title{Unmanned Aerial Vehicle R based Mission Planner  - awesome Mission Planner for complex monitoring flights
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}}
\description{
Unmanned Aerial Vehicle R based Mission Planner
}
\details{
The package provides some mission planning functionality for dealing with Unmanned Aerial Vehicles
}
\keyword{package}
