% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavRmp-package.R
\docType{package}
\name{uavRmp}
\alias{uavRmp}
\alias{uavRmp-package}
\title{UAV Mission Planner}
\description{
The package provides some mission planning functionality for dealing with Unmanned Aerial Vehicles. The focus is set on an easy to use workflow for planning autonomous obstacle avoiding surveys of rtf-UAVs to retrieve aerial or spot related data. It provides either intermediate flight control files for the DJI Litchi compatible UAVs or ready to upload control files for the pixhawk based flightcontroller as used in the Yuneec H520 or the outdated 3DR Solo. Additionally it contains some useful tools for digitizing and data manipulation.
}
\details{
The package provides some mission planning functionality for dealing with Unmanned Aerial Vehicles
}
\note{
It is important to keep in mind that all auxilliary external binaries  like GDAL or SAGA need to be installed properly.
correctly on your system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gisma/uavRmp}
  \item Report bugs at \url{https://github.com/gisma/uavRmp/issues}
}

}
\author{
Chris Reudenbach Lars Opgenoorth Sebastian Richter Florian Detsch Hanna Meyer Marvin Ludwig
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{internal}
